Fix bug when rewind to 0 position.
Bug report: The client generates a request REQ_SEEK(-1) when rewind to position 0 instead REQ_SEEK(0).
It leads to determination errors playback position because decoder_seek return -1 as new position, but -1 is a error code.

Bug fix: Not to substract 1 from the selected position at rewinding.
Bug fix(alternative): check the position of rewind and if it is <0 set it to 0.